Move globalConfigString into config file
